refactor(webgl): extract WebGL2 object-lifecycle methods (1439 → 1236 + 285)
First slice of the long-running webgl2-rendering-context.ts split.
Same `install*Methods(proto)` pattern as the parent webgl-context-base
split, applied to the WebGL2 subclass.

Extracts the five object-lifecycle method groups that share a
near-identical create/delete/is/bind shape, plus type-specific param
setters/getters:

  • VAO                — createVertexArray / deleteVertexArray /
                          isVertexArray / bindVertexArray
  • Query              — createQuery / deleteQuery / isQuery /
                          beginQuery / endQuery / getQuery /
                          getQueryParameter
  • Sampler            — createSampler / deleteSampler / isSampler /
                          bindSampler / samplerParameteri / samplerParameterf /
                          getSamplerParameter
  • Sync               — fenceSync / isSync / deleteSync / clientWaitSync /
                          waitSync / getSyncParameter
  • Transform Feedback — createTransformFeedback / ... / getTransformFeedbackVarying

32 methods totalling 203 LoC moved out of the base file.

New module:
  packages/framework/webgl/src/ts/webgl2-context/object-lifecycle.ts
  (285 LoC: interface ObjectLifecycleMethods declaration-merged into
  WebGL2RenderingContext via `declare module`, methods object, and
  the installObjectLifecycleMethods(proto) function)

Base file webgl2-rendering-context.ts 1439 → 1236 LoC. Calls
installObjectLifecycleMethods(WebGL2RenderingContext.prototype) at
the bottom plus a deliberate side-effect `import
'./webgl2-context/object-lifecycle.js'` — kept separate from the
named import so tsc preserves it in the emitted .d.ts. Same .d.ts
preservation technique as the canvas2d-core split (PR #262).

Unused `WebGLActiveInfo` import dropped from the base.

860 GJS webgl conformance tests still green; tsc + build clean.

Larger remaining method groups (texImage3D family ~190 LoC, uniform
family ~110 LoC, clearBuffer family ~110 LoC, instanced drawing +
framebuffer-ops ~150 LoC) are left in the base; future PRs can extract
them with the same pattern if motivated.

STATUS.md: new Completed entry.
