+++ STATUS.md
@@ -398,6 +398,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/webgl` — extract `WebGL2RenderingContext` object-lifecycle methods (1439 → 1236 + 285)**~~✓ — First slice of the long-running `webgl2-rendering-context.ts` split. Same `install*Methods(proto)` pattern as the parent `webgl-context-base.ts` split, applied to the WebGL2 subclass. Extracts the five object-lifecycle method groups that share a near-identical shape (`createX` / `deleteX` / `isX` / `bindX` plus type-specific param setters/getters): VAO (`createVertexArray` / `deleteVertexArray` / `isVertexArray` / `bindVertexArray`), Query (`createQuery` / `deleteQuery` / `isQuery` / `beginQuery` / `endQuery` / `getQuery` / `getQueryParameter`), Sampler (`createSampler` / `deleteSampler` / `isSampler` / `bindSampler` / `samplerParameteri` / `samplerParameterf` / `getSamplerParameter`), Sync (`fenceSync` / `isSync` / `deleteSync` / `clientWaitSync` / `waitSync` / `getSyncParameter`), Transform Feedback (`createTransformFeedback` / `deleteTransformFeedback` / `isTransformFeedback` / `bindTransformFeedback` / `beginTransformFeedback` / `endTransformFeedback` / `pauseTransformFeedback` / `resumeTransformFeedback` / `transformFeedbackVaryings` / `getTransformFeedbackVarying`). 32 methods totalling 203 LoC moved out of the base file. New module `packages/framework/webgl/src/ts/webgl2-context/object-lifecycle.ts` (285 LoC including the `interface ObjectLifecycleMethods` declaration-merged into `WebGL2RenderingContext` via `declare module`, the methods object, and the `installObjectLifecycleMethods(proto)` function). Base file `webgl2-rendering-context.ts` 1439 → 1236 LoC, calls `installObjectLifecycleMethods(WebGL2RenderingContext.prototype)` after class declaration plus a deliberate side-effect `import './webgl2-context/object-lifecycle.js'` retained in the emitted `.d.ts` so downstream consumers see the augmentations. Same `.d.ts` preservation technique as the canvas2d-core split (PR #262). Unused `WebGLActiveInfo` import dropped from the base. 860 GJS webgl conformance tests still green. Larger remaining method groups (texImage3D family ~190 LoC, uniform family ~110 LoC, clearBuffer family ~110 LoC, instanced drawing + framebuffer-ops ~150 LoC) are left in the base; future PRs can extract them with the same pattern if motivated.
+
 - ~~**`as any` cleanup batch — impl-code edition (48 → 0 real casts across 7 files)**~~✓ — Fourth batch in the `as any` reduction sweep, this time targeting actual implementation code (not register modules). Each fix surfaces real type-safety where the previous `as any` was eating an assignability check. Files touched (before → after, all 0 real casts remaining, residual matches are doc comments): `packages/web/webrtc/src/get-user-media.ts` 9 → 0 (new `interface _GstElementProps extends Gst.Element` capturing the dynamic GObject properties `caps` / `is_live` / `wave` / `pattern` that don't appear on the GIR-typed `Gst.Element`), `packages/web/xmlhttprequest/src/index.ts` 8 → 0 (new `interface _URLObjectURLPatch` for the `URL.createObjectURL` / `URL.revokeObjectURL` / `__gjsify_objecturl` properties this module patches onto the URL constructor; `(this as any)['on' + type]` for XHR event handlers becomes a typed `Record<\`on${XHREventType}\`, …>` lookup), `packages/framework/video/src/pipeline-builder.ts` 7 → 0 (local `interface _PaintableSink extends Gst.Element` / `interface _GlSinkBin extends Gst.Element` / `interface _Playbin extends Gst.Element` for `gtk4paintablesink`/`glsinkbin`/`playbin` GObject properties; `request_pad_simple` vs legacy `get_request_pad` probed via a typed-optional structural fallback), `packages/web/webrtc/src/register/peer-connection.ts` (handled in earlier register PR — not this one), `packages/node/sqlite/src/param-binding.ts` 6 → 0 (new `setHolderPrim(holder, value)` helper consolidating the `holder.set_value(prim as unknown as GObject.Value)` cast so each call site stays free of `any`; this hides GJS' auto-conversion from JS primitives to `GObject.Value` behind one place), `packages/node/perf_hooks/src/index.ts` 6 → 0 (new `interface _PerfHooksGlobals` for the optional native `PerformanceObserver` / `PerformanceEntry` / `PerformanceObserverEntryList` / `PerformanceMark` / `PerformanceMeasure` constructors; the GLib bootstrap path also gains a structural `(globalThis as {imports?:{gi?:{GLib?:{get_monotonic_time():number}}}})` view), `packages/node/fs/src/fd-ops.ts` 6 → 0 (`fstatAsync` now picks the right `fstatSync` overload at runtime based on `options?.bigint`; `readSync`/`writeSync` use the inferred narrow types instead of casting each parameter to `any`), `packages/framework/webgl/src/ts/webgl-bridge.ts` 6 → 0 (local `interface _RafGlobals` for `requestAnimationFrame`/`cancelAnimationFrame`/`performance` and a typed `{document?:{body?:{appendChild(el:unknown):void}}}` view for the GTK→DOM document attach). tsc clean across all 7 touched packages. Net change: 48 real `as any` casts removed; each replacement is more verbose by 5-10 LoC (the typed-view declaration), so total LoC grew by ~80 — but every call site now type-checks the property access at compile time. The `(this as any)['on' + type]` change in xmlhttprequest in particular surfaces a real protection: if a future PR adds a new `XHREventType` member it must also add the matching `on<X>` property to the class signature, otherwise the typed lookup will fail.
 
 - ~~**`as any` cleanup batch — register modules (53 → 5 real casts)**~~✓ — Third batch in the long-running `as any` reduction sweep, this time targeting the densest hotspots in `/register` side-effect modules across the web + DOM pillars. Same approach as PR #263: each module now has a single module-local `interface _<Name>Globals` typed view of the `globalThis` slots it writes (and reads for the existence-guard), declared once at the top and used through the install branch. Files touched (before → after): `packages/web/streams/src/register/readable.ts` 12 → 1 (the kept cast is the typed-view itself), `packages/web/webrtc/src/register/peer-connection.ts` 8 → 0, `packages/web/webaudio/src/register.ts` 8 → 0, `packages/web/web-globals/src/register.ts` 6 → 1, `packages/web/webrtc/src/register/media.ts` 6 → 0, `packages/web/websocket/src/register.ts` 6 → 0, `packages/dom/dom-elements/src/register/document.ts` 7 → 1. The dom-elements register file additionally needed `Parameters<EventTarget['addEventListener']>[…]` indirection to keep the global-side listener / Event types assignable to `@gjsify/dom-events`'s structurally-narrower ones — a real type-safety net that the previous `as any` cast was eating silently. tsc clean across all 6 packages. Total real cast reduction: 53 → 5 (the 5 remaining are the typed-view `as unknown as` lines themselves, no `any` involved).
