+++ packages/framework/webgl/src/ts/webgl2-rendering-context.ts
@@ -12,7 +12,6 @@ import { WebGLSync } from './webgl-sync.js';
 import { WebGLTransformFeedback } from './webgl-transform-feedback.js';
 import { WebGLVertexArrayObject } from './webgl-vertex-array-object.js';
 import { WebGLUniformLocation } from './webgl-uniform-location.js';
-import { WebGLActiveInfo } from './webgl-active-info.js';
 import { WebGLProgram as OurWebGLProgram } from './webgl-program.js';
 import { WebGLBuffer as OurWebGLBuffer } from './webgl-buffer.js';
 import { WebGLTexture } from './webgl-texture.js';
@@ -425,209 +424,6 @@ export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2Re
         }
     }
 
-    // ─── Vertex Array Objects ─────────────────────────────────────────────
-
-    createVertexArray(): WebGLVertexArrayObject | null {
-        const id = this._native2.createVertexArray();
-        if (!id) return null;
-        const vao = new WebGLVertexArrayObject(id, this);
-        this._vertexArrayObjects[id] = vao;
-        return vao;
-    }
-
-    deleteVertexArray(vertexArray: WebGLVertexArrayObject | null): void {
-        if (!vertexArray || !(vertexArray instanceof WebGLVertexArrayObject)) return;
-        vertexArray._pendingDelete = true;
-        vertexArray._checkDelete();
-    }
-
-    isVertexArray(vertexArray: WebGLVertexArrayObject | null): GLboolean {
-        if (!vertexArray || !(vertexArray instanceof WebGLVertexArrayObject)) return false;
-        return this._native2.isVertexArray(vertexArray._);
-    }
-
-    bindVertexArray(array: WebGLVertexArrayObject | null): void {
-        if (array === null) {
-            this._native2.bindVertexArray(0);
-            this._vertexObjectState = this._defaultVertexObjectState;
-        } else if (array instanceof WebGLVertexArrayObject) {
-            this._native2.bindVertexArray(array._);
-            this._vertexObjectState = array._objectState;
-        } else {
-            this.setError(this.INVALID_OPERATION);
-        }
-    }
-
-    // ─── Query Objects ────────────────────────────────────────────────────
-
-    createQuery(): WebGLQuery | null {
-        const id = this._native2.createQuery();
-        if (!id) return null;
-        const query = new WebGLQuery(id, this);
-        this._queries[id] = query;
-        return query;
-    }
-
-    deleteQuery(query: WebGLQuery | null): void {
-        if (!query || !(query instanceof WebGLQuery)) return;
-        query._pendingDelete = true;
-        query._checkDelete();
-    }
-
-    isQuery(query: WebGLQuery | null): GLboolean {
-        if (!query || !(query instanceof WebGLQuery)) return false;
-        return this._native2.isQuery(query._);
-    }
-
-    beginQuery(target: GLenum, query: WebGLQuery): void {
-        if (!(query instanceof WebGLQuery)) return;
-        this._native2.beginQuery(target, query._);
-    }
-
-    endQuery(target: GLenum): void {
-        this._native2.endQuery(target);
-    }
-
-    getQuery(_target: GLenum, _pname: GLenum): WebGLQuery | null {
-        warnNotImplemented('WebGL2RenderingContext.getQuery');
-        return null;
-    }
-
-    getQueryParameter(query: WebGLQuery, pname: GLenum): any {
-        if (!(query instanceof WebGLQuery)) return null;
-        return this._native2.getQueryParameter(query._, pname);
-    }
-
-    // ─── Sampler Objects ──────────────────────────────────────────────────
-
-    createSampler(): WebGLSampler | null {
-        const id = this._native2.createSampler();
-        if (!id) return null;
-        const sampler = new WebGLSampler(id, this);
-        this._samplers[id] = sampler;
-        return sampler;
-    }
-
-    deleteSampler(sampler: WebGLSampler | null): void {
-        if (!sampler || !(sampler instanceof WebGLSampler)) return;
-        sampler._pendingDelete = true;
-        sampler._checkDelete();
-    }
-
-    isSampler(sampler: WebGLSampler | null): GLboolean {
-        if (!sampler || !(sampler instanceof WebGLSampler)) return false;
-        return this._native2.isSampler(sampler._);
-    }
-
-    bindSampler(unit: GLuint, sampler: WebGLSampler | null): void {
-        this._native2.bindSampler(unit, sampler ? sampler._ : 0);
-    }
-
-    samplerParameteri(sampler: WebGLSampler, pname: GLenum, param: GLint): void {
-        if (!(sampler instanceof WebGLSampler)) return;
-        this._native2.samplerParameteri(sampler._, pname, param);
-    }
-
-    samplerParameterf(sampler: WebGLSampler, pname: GLenum, param: GLfloat): void {
-        if (!(sampler instanceof WebGLSampler)) return;
-        this._native2.samplerParameterf(sampler._, pname, param);
-    }
-
-    getSamplerParameter(sampler: WebGLSampler, pname: GLenum): any {
-        if (!(sampler instanceof WebGLSampler)) return null;
-        // Float params: TEXTURE_MIN_LOD, TEXTURE_MAX_LOD
-        if (pname === 0x813A || pname === 0x813B) {
-            return this._native2.getSamplerParameterf(sampler._, pname);
-        }
-        return this._native2.getSamplerParameteri(sampler._, pname);
-    }
-
-    // ─── Sync Objects ─────────────────────────────────────────────────────
-
-    fenceSync(condition: GLenum, flags: GLbitfield): WebGLSync | null {
-        const id = this._native2.fenceSync(condition, flags);
-        if (!id) return null;
-        const sync = new WebGLSync(id, this);
-        this._syncs[id] = sync;
-        return sync;
-    }
-
-    isSync(sync: WebGLSync | null): GLboolean {
-        if (!sync || !(sync instanceof WebGLSync)) return false;
-        return this._native2.isSync(sync._);
-    }
-
-    deleteSync(sync: WebGLSync | null): void {
-        if (!sync || !(sync instanceof WebGLSync)) return;
-        sync._pendingDelete = true;
-        sync._checkDelete();
-    }
-
-    clientWaitSync(sync: WebGLSync, flags: GLbitfield, timeout: GLuint64): GLenum {
-        if (!(sync instanceof WebGLSync)) return 0x911C; // WAIT_FAILED
-        return this._native2.clientWaitSync(sync._, flags, timeout as unknown as number);
-    }
-
-    waitSync(sync: WebGLSync, flags: GLbitfield, timeout: GLint64): void {
-        if (!(sync instanceof WebGLSync)) return;
-        this._native2.waitSync(sync._, flags, timeout as unknown as number);
-    }
-
-    getSyncParameter(sync: WebGLSync, pname: GLenum): any {
-        if (!(sync instanceof WebGLSync)) return null;
-        return this._native2.getSyncParameter(sync._, pname);
-    }
-
-    // ─── Transform Feedback ───────────────────────────────────────────────
-
-    createTransformFeedback(): WebGLTransformFeedback | null {
-        const id = this._native2.createTransformFeedback();
-        if (!id) return null;
-        const tf = new WebGLTransformFeedback(id, this);
-        this._transformFeedbacks[id] = tf;
-        return tf;
-    }
-
-    deleteTransformFeedback(tf: WebGLTransformFeedback | null): void {
-        if (!tf || !(tf instanceof WebGLTransformFeedback)) return;
-        tf._pendingDelete = true;
-        tf._checkDelete();
-    }
-
-    isTransformFeedback(tf: WebGLTransformFeedback | null): GLboolean {
-        if (!tf || !(tf instanceof WebGLTransformFeedback)) return false;
-        return this._native2.isTransformFeedback(tf._);
-    }
-
-    bindTransformFeedback(target: GLenum, tf: WebGLTransformFeedback | null): void {
-        this._native2.bindTransformFeedback(target, tf ? tf._ : 0);
-    }
-
-    beginTransformFeedback(primitiveMode: GLenum): void {
-        this._native2.beginTransformFeedback(primitiveMode);
-    }
-
-    endTransformFeedback(): void {
-        this._native2.endTransformFeedback();
-    }
-
-    pauseTransformFeedback(): void {
-        this._native2.pauseTransformFeedback();
-    }
-
-    resumeTransformFeedback(): void {
-        this._native2.resumeTransformFeedback();
-    }
-
-    transformFeedbackVaryings(program: WebGLProgram, varyings: string[], bufferMode: GLenum): void {
-        this._native2.transformFeedbackVaryings((program as unknown as OurWebGLProgram)._, varyings, bufferMode);
-    }
-
-    getTransformFeedbackVarying(program: WebGLProgram, index: GLuint): WebGLActiveInfo | null {
-        const result = this._native2.getTransformFeedbackVarying((program as unknown as OurWebGLProgram)._, index)
-            .deepUnpack<{ name: string; size: number; type: number }>();
-        return new WebGLActiveInfo({ size: result.size, type: result.type, name: result.name });
-    }
 
     // ─── Indexed Buffer Binding ───────────────────────────────────────────
 
@@ -1437,3 +1233,13 @@ export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2Re
         return true;
     }
 }
+
+// Wire focused method groups into WebGL2RenderingContext.prototype, same
+// pattern as the parent `WebGLContextBase` split. The side-effect import is
+// kept separate from the named import so tsc preserves it in the emitted
+// `.d.ts` — downstream consumers need the `declare module` augmentations
+// loaded to see VAO / Query / Sampler / Sync / TransformFeedback methods on
+// the published type.
+import './webgl2-context/object-lifecycle.js';
+import { installObjectLifecycleMethods } from './webgl2-context/object-lifecycle.js';
+installObjectLifecycleMethods(WebGL2RenderingContext.prototype);
